pyproject.toml
[tool.poetry]
name = "barcode-verification"
version = "0.1.0"
description = "Daejin barcode verification program"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.0"
python-dotenv = "^0.19.0"
tkinter = "*"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
black = "^22.0.0"
isort = "^5.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
